ner/requirements.txt
pytorch_pretrained_bert>=0.4.0
mxnet>=1.5.0
gluonnlp>=0.6.0
sentencepiece>=0.1.6
git+https://github.com/kmkurn/pytorch-crf#egg=pytorch_crf
transformers==2.6.0
tb-nightly
future
tensorflow
konlpy
tqdm
hgtk
nltk

# apt install default-jre (for konlpy)